requirements.txt
# Core AI and ML libraries
torch>=2.0.0
transformers>=4.30.0
whisper-openai>=20231117
numpy>=1.24.0
scipy>=1.10.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
pyaudio>=0.2.11

# Text processing and embeddings
sentence-transformers>=2.2.0
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Computer vision (optional, for emotion detection)
opencv-python>=4.8.0
mediapipe>=0.10.0

# Text-to-Speech
pyttsx3>=2.90
gTTS>=2.3.0

# Database and storage
sqlite3
cryptography>=41.0.0

# Web interface (optional)
flask>=2.3.0
flask-cors>=4.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
click>=8.1.0
tqdm>=4.65.0
python-dateutil>=2.8.2

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
